Fixed narrowing error  (converting int to double)
changed Physics_rect to have all doubles this caused a narrowing error in the constructor of the RigidBody class so I fixed that too

+++ RPG_Lib/Physics/Physics.hpp
@@ -10,7 +10,7 @@ namespace Skele_lib {
 				movable objectSpeed = { 0,0,0,0 };
 				movable drag = { 2,0,0,0 };
 			public:
-				RigidBody(int x, int y, float w, float h) {
+				RigidBody(double x, double y, double w, double h) {
 					rect = { x,y,w,h }; 
 				}
 				Physics_rect getPosition() {
